requirements.txt
asgiref==3.5.2
boto3==1.24.92
botocore==1.27.92
Django==4.1.2
django-appconf==1.0.5
django-filter==22.1
django-imagekit==4.1.0
django-storages==1.13.1
jmespath==1.0.1
pilkit==2.0
Pillow==9.2.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
